google-pub-sub-example/requirements.txt
google-cloud
google-cloud-pubsub
google-cloud-logging
